fix: put aria sort on th not button
